% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\docType{methods}
\name{abstractionLayer,decisionMatrix-method}
\alias{abstractionLayer,decisionMatrix-method}
\title{Extract abstraction layer from descisionMatrix object}
\usage{
\S4method{abstractionLayer}{decisionMatrix}(object)
}
\arguments{
\item{object}{decisionMatrix.}
}
\value{
list describing relationship between chromatin marks and functional categories
}
\description{
Extract abstraction layer from descisionMatrix object
}

