% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cedars.BFG.colors}
\alias{Cedars.BFG.colors}
\title{Default Segmentation Colors}
\format{A data.frame with 20 states and colors:
\describe{
  \item{TRS}{Transcribed Regions}
  \item{HET}{Heterochromatin Regions}
  \item{SCR}{Silenced Regions}
  \item{PAR}{Promotor Active Regions}
  \item{PARC}{Promotor Active Regions, Core}
  \item{PPR}{Promotor Poised Regions}
  \item{PPRC}{Promotor Poised Regions, Core}
  \item{EAR}{Enhancer Active Regions}
  \item{EARC}{Enhancer Active Regions, Core}
  \item{EPR}{Enhancer Poised Regions}
  \item{EPRC}{Enhancer Poised Regions, Core}
  \item{ER}{Enhancer Regions}
  \item{ERC}{Enhancer Regions, Core}
  \item{PR}{Promoter Regions}
  \item{PRC}{Promoter Regions, Core}
  \item{CTCF}{CTCF Regions}
  \item{CTCFC}{CTCF Regions, Core}
  \item{RPS}{Regulatory Putative Region (DNase1 Hypersensitivity, ATAC-Seq, or TF binding)}
  \item{AR}{Active Region}
  \item{ARC}{Active Region, Core}
  }}
\source{
\url{"http://github.com/Simon-Coetzee/StateHub/"}
}
\usage{
Cedars.BFG.colors
}
\value{
\code{matrix} of states
}
\description{
A table describing the default colors for the default
state definitions
}
\details{
load with \code{data(Cedars.BFG.states)}

Colors are intended to be simmilar to chromHMM to facilitate comparison.
}
\examples{
data(Cedars.BFG.colors)
Cedars.BFG.colors
}
\keyword{datasets}

