% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cedars.BFG.states}
\alias{Cedars.BFG.states}
\title{Default Segmentation Definitions}
\format{An object of class \code{matrix} with 20 rows and 8 columns.}
\source{
\url{"http://github.com/Simon-Coetzee/StateHub/"}
}
\usage{
Cedars.BFG.states
}
\description{
A table describing the default state definitions used with the default
translation layer
}
\keyword{datasets}

