% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\docType{class}
\name{decisionMatrix-class}
\alias{decisionMatrix-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{id}}{character.}

\item{\code{name}}{character.}

\item{\code{author}}{character.}

\item{\code{revision}}{character.}

\item{\code{description}}{character.}

\item{\code{abstraction.layer}}{list.}

\item{\code{decision.matrix}}{matrix.}

\item{\code{state.colors}}{character.}
}}

