% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\docType{methods}
\name{decisionMatrix,decisionMatrix-method}
\alias{decisionMatrix,decisionMatrix-method}
\title{Extract Descision Matrix from descisionMatrix object}
\usage{
\S4method{decisionMatrix}{decisionMatrix}(object)
}
\arguments{
\item{object}{decisionMatrix.}
}
\value{
matrix descisionMatrix stripped of metadata
}
\description{
Extract Descision Matrix from descisionMatrix object
}

