% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\name{ExportStatePaintR}
\alias{ExportStatePaintR}
\title{Write StatePaintR object to bedfiles}
\usage{
ExportStatePaintR(states, color.key = Cedars.BFG.colors,
  output.dir = tempdir())
}
\description{
Write StatePaintR object to bedfiles
}

