% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint.R
\name{getFootprint}
\alias{getFootprint}
\title{getFootprint}
\usage{
getFootprint(directory = NULL, celltype = NULL, deftable = NULL,
  deflookup = NULL)
}
\arguments{
\item{deftable}{}
}
\description{
getFootprint
}

