% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\name{PaintStates}
\alias{PaintStates}
\title{Run a StatePaintR decisionMatrix against a group of files listed in a manifest.}
\usage{
PaintStates(manifest, decisionMatrix, scoreStates = FALSE, progress = TRUE)
}
\arguments{
\item{progress}{}
}
\description{
Run a StatePaintR decisionMatrix against a group of files listed in a manifest.
}

