% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\name{PlotStates}
\alias{PlotStates}
\title{PlotStates - plotting the results of PaintStates}
\usage{
PlotStates(states, location = NULL, gene.track = NULL,
  additional.tracks = NULL)
}
\arguments{
\item{states}{A GRangesList produced by \code{\link{PaintStates}} or a list
of GRanges, or GRangesList that has a "score" column, and optionally a
"state" column, that is filtered on if the \code{state.select} argument is
used}

\item{location}{A GRanges object of length 1, or a character vector of the
form: "chr:start-end", e.g., "chr12:51267156-52080611"}

\item{gene.track}{optionally, an object created by BiomartGeneRegionTrack() to plot the gene context in addition to segmentation.}

\item{additional.tracks}{optionally, an object of class "AnnotationTrack" with additional data to plot in addition to segmentation.}
}
\value{
Plots the states, of the samples in the states object, and if possible the genes in the region
}
\description{
PlotStates - plotting the results of PaintStates
}
\examples{
load(system.file("extdata", "heart.states.rda", package = "StatePaintR"))
load(system.file("extdata", "vista.heart.enhancers.rda", package = "StatePaintR"))
mylocation <- GRanges("chr1:42623508-42754885")
vista.in.range <- subsetByOverlaps(heart.enhancers, mylocation)
vista.anno <- AnnotationTrack(range = vista.in.range,
                              fill = mcols(vista.in.range)$validated + 2,
                              stacking = "dense",
                              col = NULL,
                              col.line = NULL,
                              stackHeight = 1,
                              shape = "box",
                              rotation.title = 360,
                              background.title = "transparent", col.title = "black",
                              cex.title = 0.5,
                              name = "VISTA enhancers")
PlotStates(states = heart.states,
           location = mylocation,
           additional.tracks = vista.anno)
}
