% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prg.R
\name{prg_convex_hull}
\alias{prg_convex_hull}
\title{Create the convex hull of the Precision-Recall-Gain curve}
\usage{
prg_convex_hull(prg_curve)
}
\arguments{
\item{prg_curve}{the data structure resulting from the function
create_prg_curve}
}
\value{
the data.frame representing the convex hull
}
\description{
This function creates the convex hull of the Precision-Recall-Gain curve
resulting from the function create_prg_curve and calculates the F-calibrated
scores. More information on Precision-Recall-Gain curves and how to cite this
work is available at http://www.cs.bris.ac.uk/~flach/PRGcurves/.
}
