% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prg.R
\name{recall_gain}
\alias{recall_gain}
\title{Recall Gain}
\usage{
recall_gain(TP, FN, FP, TN)
}
\arguments{
\item{TP}{number of true positives, can be a vector}

\item{FN}{number of false negatives, can be a vector}

\item{FP}{number of false positives, can be a vector}

\item{TN}{number of true negatives, can be a vector}
}
\value{
Recall Gain (a numeric value less than or equal to 1; or -Inf or NaN, see the details below)
}
\description{
This function calculates Recall Gain from the entries of the contingency table: number of true positives (TP), false negatives (FN), false positives (FP), and true negatives (TN). More information on Precision-Recall-Gain curves and how to cite this work is available at http://www.cs.bris.ac.uk/~flach/PRGcurves/.
}
\details{
Recall Gain (RecGain) quantifies by how much recall is improved over the recall equal to the proportion of positives (pi). RecGain=1 stands for maximal improvement (Rec=1) and RecGain=0 stands for no improvement (Rec=pi). If Rec=0, then RecGain=-Inf. It can happen that RecGain=NaN, for instance if there are no negatives (FP=0 and TN=0) and FN>0 and TP=0.
}
\examples{
recall_gain(3,0,1,2)
# [1] 1
TP = c(0,2,3)
FN = 3-TP
FP = c(0,1,2)
TN = 2-FP
recall_gain(TP,FN,FP,TN)
# [1]  -Inf 0.25 1.00
}
