% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\docType{methods}
\name{stateColors,decisionMatrix-method}
\alias{stateColors,decisionMatrix-method}
\title{Extract color information from descisionMatrix object}
\usage{
\S4method{stateColors}{decisionMatrix}(object)
}
\arguments{
\item{object}{decisionMatrix.}
}
\value{
named vector of colors for chromatin states
}
\description{
Extract color information from descisionMatrix object
}

