% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\name{stateColors}
\alias{stateColors}
\title{stateColors}
\usage{
stateColors(object)
}
\arguments{
\item{object}{of class decisionMatrix}
}
\value{
character. The color to assign to each state.
}
\description{
stateColors
}
\examples{
load(system.file("extdata", "poised.promoter.model.rda", package = "StatePaintR"))
poised.promoter.model
abstractionLayer(poised.promoter.model)
}
