% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\name{ExportStatePaintR}
\alias{ExportStatePaintR}
\title{Write StatePaintR object to bedfiles}
\usage{
ExportStatePaintR(states, output.dir, progress = TRUE)
}
\arguments{
\item{states}{A GRangesList produced by \code{\link{PaintStates}}}

\item{output.dir}{A character string indicating the directory to save the
exported states. The directory will be created if it does not exist.}

\item{progress}{logical; show progress bar and timing details?}
}
\value{
Invisibly returns the states object.
}
\description{
Write StatePaintR object to bedfiles
}
\examples{
\dontrun{
ExportStatePaintR(states = states,
                  output.dir = tempdir())
}
}
