% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\name{get.decision.matrix}
\alias{get.decision.matrix}
\title{Retrieve Decision Matrix from StateHub}
\usage{
get.decision.matrix(search)
}
\arguments{
\item{search}{character string of either unique id, or search term.}
}
\value{
decisionMatrix object
}
\description{
Retrieve Decision Matrix from StateHub
}

