% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint.R
\name{getFootprint}
\alias{getFootprint}
\title{getFootprint}
\usage{
getFootprint(directory = NULL, search = NULL, chrome_states = "default",
  translation_layer = "default")
}
\arguments{
\item{translation_layer}{}
}
\description{
getFootprint
}

