% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepaintr.R
\name{PaintStates}
\alias{PaintStates}
\title{getFootprint}
\usage{
PaintStates(manifest = NULL, chrome_states = "default",
  translation_layer = "default")
}
\arguments{
\item{translation_layer}{}
}
\description{
getFootprint
}

